project_rag/requirements.txt
fastapi>=0.109.2,<1.0.0
uvicorn>=0.27.1,<1.0.0
sqlalchemy>=2.0.0,<3.0.0
jinja2>=3.1.3,<4.0.0
alembic>=1.13.1,<2.0.0
aiosqlite>=0.19.0,<1.0.0
pydantic-settings>=2.2.0,<3.0.0

# New requirements
llama-index>=0.10.0,<0.11.0
llama-index-embeddings-huggingface>=0.1.4,<0.2.0
llama-index-llms-llama-cpp>=0.1.3,<0.2.0
llama-index-llms-together>=0.1.3,<0.2.0
python-dotenv